1st Commit
Create a Python calculator program that performs basic arithmetic operations: addition, subtraction, multiplication, and division. 

Input
Ensure that the user inputs valid numbers (for the First number and second Number)
Verify that the operation entered is one of the following: +, -, *, or /. 

Functionality
Prompt the user to enter two numbers and select an arithmetic operation.
Perform the chosen operation and display the result.
After each calculation, ask the user if they want to perform another calculation or exit.
If the user wants to continue, repeat the process.
If the user chooses to exit, end the program.